+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Completed unit test coverage for C# generation.
 - RequiredAttribute is now being generated.
 - Updated examples.
+- Added validation for entities, enums and type aliases identifiers
 
 ## Version 0.1
 - Changed main artifact name to include '-release' suffix.